### 2023-06-29 #595 @katarzynam-165
rangr: Species range dynamics simulation toolset
Submitting Author Name: Katarzyna Markowska
Submitting Author Github Handle: @katarzynam-165
Other Package Authors Github handles: @LechoslawKuczynski
Repository: https://github.com/popecol/rangr
Version submitted: 1.0.0
Submission type: Stats
Badge grade: silver
Reviewers: taddallas, @taddallas, @TheAnalyticalEdge
Due date for taddallas: 2023-08-16
Version accepted: TBD
---

```
Package: rangr
Title: Mechanistic Simulation of Species Range Dynamics
Version: 1.0.0
    person("Katarzyna", "Markowska", email = "katarzyna.markowska@amu.edu.pl", role = c("aut", "cre")),
    person("Lechosław", "Kuczyński", email = "lechu@amu.edu.pl", role = "aut"))
Description: Species range dynamics simulation toolset.
    parallel,
    pbapply,
    grDevices,
    graphics,
    zoo,
    terra,
    raster,
    assertthat
    covr,
    bookdown
Roxygen: list (markdown = TRUE, roclets = c ("namespace", "rd", "srr::srr_stats_roclet"))
    https://popecol.github.io/rangr/
```

-   Please indicate which of our [statistical package categories](https://stats-devguide.ropensci.org/overview.html#overview-categories) this package falls under. (Please check one appropriate box below):

-   [x] A pre-submission inquiry has been approved in [issue#\<592\>](#%3C592%3E)

    rangr is an R package designed for simulating species range dynamics, primarily aimed at ecologists and conservationists who work with complex data structures such as those derived from citizen science and wildlife monitoring programs. With rangr, users can mimic the key processes that shape population numbers and spatial distributions, including local dynamics, dispersal, and habitat selection, to project population responses to environmental changes. Additionally, rangr can be used to test and evaluate different methods of modelling species distribution using simulated data as a reference.

    -   *The first implementation of a novel algorithm*; or
    -   *The first implementation within* **R** *of an algorithm which has previously been implemented in other languages or contexts*; or
    -   *An improvement on other implementations of similar algorithms in* **R**.

    Please include hyperlinked references to all other relevant software.

    rangr is the first implementation of a novel algorithm, but there are a few packages like [RangeShiftR](https://github.com/RangeShifter/RangeShiftR-package), [poems](https://github.com/GlobalEcologyLab/poems), or [steps](https://github.com/steps-dev/steps) that serve similar purposes. However, none of them met all the criteria that were important to us in this type of simulation, such as being easy to set up and customize with other existing R functions, supporting simulations that vary in both time and space, and incorporating the Virtual Ecologist approach by providing functions for various sampling scenarios.

    Silver

    > Have a demonstrated *generality* of usage beyond one single envisioned use case. Software is frequently developed for one particular use case envisioned by the authors themselves. Generalising the utility of software so that it is readily applicable to other use cases, and satisfactorily documenting such generality of usage, represents another aspect which may be considered sufficient for software to attain a silver grade.

    This aspect is particularly well-suited due to the versatile range of applications offered by rangr. These applications include:

    -   modelling population range dynamics,

    -   testing various ecological scenarios,

    -   testing species distribution modelling methods.
