### 2023-07-10 #595 @adamhsparks
@katarzynam-165, can you address my editor comments and get the ‘srr’ standards compliance up? Perhaps there is an easy standard that you can add or meet?
