### 2023-07-10 #595 @katarzynam-165
@adamhsparks thanks for the feedback! I will make the requested changes and review the NA standards - I'm pretty sure I can meet or add at least one (I was so close the first time :( ). I'll get back to you soon.
