### 2023-07-16 4abd746
Create CODE_OF_CONDUCT.md
