### 2023-07-18 e616e15
Minor changes to the vignette
