+++ vignettes/rangr.Rmd
@@ -141,11 +141,11 @@ Let's break down this command:
 
 -   The `r` parameter is used to set the intrinsic population growth rate. The default population growth function is the Gompetz function.
 
--   The `rate` parameter is related to the `kernel_fun` parameter, which by default is set as an exponential function (`rexp`). Therefore, `rate` determines the shape of the dispersal function.
+-   The `rate` parameter is related to the `kernel_fun` parameter, which by default is set to an exponential function (`rexp`). Therefore, `rate` determines the shape of the dispersal kernel (the mean dispersal distance is 1/rate).
 
-That way we have determined simulation environment along with demographic processes and dispersal. That’s the most basic set-up needed to perform your first simulation and we will do just that in the next step.
+We have now set up the simulation environment along with the demographic and dispersal processes. This is the most basic setup needed to run your first simulation, and that's what we'll do in the next step.
 
-But before that, let’s see what other information `sim_data` contains and what happened behind the scene during initialisation.
+But first, let's see what other information `sim_data` contains and what happened behind the scenes during initialisation.
 
 First, let’s check the class of `sim_data` object:
 
@@ -165,9 +165,9 @@ It will show the summary of both input maps, as well as list of other most impor
 
 ## Simulation
 
-All you need to perform a simulation is a `sim_data` object and a specified number of time steps that you want to simulate. Additionally, you can use the `burn` parameter to discard a selected number of initial time steps if it makes sense for your research or experiment.
+All you need to run a simulation is a `sim_data` object and a specified number of time steps to simulate. In addition, you can use the `burn` parameter to discard a selected number of initial time steps if this makes sense for your research or experiment.
 
-In this first example, we will only set the `time` parameter since we want to observe how our virtual species is dispersing and reproducing.
+In this first example, we will only set the `time` parameter as we want to observe how our virtual species disperses and reproduces.
 
 ```{r sim_result_01, eval=FALSE}
 sim_result_01 <- sim(obj = sim_data_01, time = 100)
@@ -181,11 +181,11 @@ class(sim_result_01)
 
 Similar to `initialise`, `sim` also returns an object that inherits from `list`, but in this case it is called `sim_results`. This list has 3 elements such as:
 
--   `extinction` - `TRUE` if population is extinct or `FALSE` otherwise
+-   `extinction` - `TRUE` if the population is extinct or `FALSE` otherwise
 
--   `simulated_time` - number of simulated time steps without the burn-in ones
+-   `simulated_time` - number of simulated time steps without the burn-in
 
--   `N_map` - 3-dimensional array representing spatiotemporal variability in population numbers. The first two dimensions correspond to the spatial aspect of the output and the third dimension represents time.
+-   `N_map` - 3-dimensional array representing the spatiotemporal variability of population numbers. The first two dimensions correspond to the spatial aspect of the output and the third dimension represents time.
 
 The best way to take a closer look at the results is to call a summary function.
 
@@ -193,11 +193,11 @@ The best way to take a closer look at the results is to call a summary function.
 summary(sim_result_01)
 ```
 
-It gives you a quick and easy overview of the simulation results by providing simulation time, extinction status and summary of all maps with abundances. It also produces a plot that can be useful to determine the value for `burn` parameter.
+It gives you a quick and easy overview of the simulation results by providing simulation time, extinction status and a summary of all maps with abundances. It also produces a plot which can be useful for determining the value of the `burn` parameter.
 
 ## Visualisation
 
[…]
