### 2023-07-18 #595 @katarzynam-165
Hi @adamhsparks!
I have made the requested changes and would like to thank you especially for the helpful comment regarding the "Reference" page. It looks much clearer now.
Additionally, I have implemented some standards that should improve compliance with 'srr' standards.
If there are any other improvements that you think should be made to the package, please let me know :)
