### 2023-07-19 #595 @adamhsparks
Thank you, @katarzynam-165. You're right, the reference page is now much more easily read and understood. I appreciate you taking the time to address these issues.

I'll start looking for reviewers now.
