### 2023-07-31 #595 @taddallas
## Package Review

-   ☒ As the reviewer I confirm that there are no [conflicts of interest](https://devguide.ropensci.org/policies.html#coi) for me to review this work (if you are unsure whether you are in conflict, please speak to your editor *before* starting your review).

#### Documentation

-   ☒ **A statement of need:** clearly stating problems the software is designed to solve and its target audience in README
-   ☒ **Installation instructions:** for the development version of package and any non-standard dependencies in README
-   ☒ **Vignette(s):** demonstrating major functionality that runs successfully locally
-   ☒ **Function Documentation:** for all exported functions
-   ☒ **Examples:** (that run successfully locally) for all exported functions
-   ☒ **Community guidelines:** including contribution guidelines in the README or CONTRIBUTING, and DESCRIPTION with `URL`, `BugReports` and `Maintainer` (which may be autogenerated via `Authors@R`).

#### Functionality

-   ☒ **Installation:** Installation succeeds as documented.
-   ☒ **Functionality:** Any functional claims of the software been confirmed.
-   ☒ **Performance:** Any performance claims of the software been confirmed.
-   ☒ **Automated tests:** Unit tests cover essential functions of the package and a reasonable range of inputs and conditions. All tests pass on the local machine.
-   ☒ **Packaging guidelines**: The package conforms to the rOpenSci packaging guidelines.

Estimated hours spent reviewing: 3-4

-   ☒ Should the author(s) deem it appropriate, I agree to be acknowledged as a package reviewer (“rev” role) in the package DESCRIPTION file.

### Review Comments

I get a warning when loading the `sp` package (it is a dependency of `raster`). Have the authors developed a plan for handling this?

> The legacy packages maptools, rgdal, and rgeos, underpinning the sp package, which was just loaded, will retire in October 2023.

Consider adding the `Code_of_Conduct.md` to `.Rbuildignore` (it gets detected as a non-standard file when checking the package using `devtools`).

Does the `initalize` function have default values for some of the arguments like `dens_dep` and `border`? It seems like ‘absorbing’ is the default for `border` and ‘K2N’ is default for `dens_dep`, but this should maybe be noted in the documentation?

A small note, but the authors argue that `r_sd` is demographic stochasticity, but it is a time-varying growth rate applied to all cells equally (this may or may not be right). This seems like a measure of environmental stochasticity (e.g., there are good years and bad years). Perhaps future development of the package could include spatial variation in growth rates? Covariation between r and K (or the variance therein)? One way the authors could do this is to have the `r` argument take any number of forms (a raster the same size as `n`, a single value, etc.). It’s tough to do the change through time bits, as the `initialize` step sets up so much of the hyperparameters, but the simulation is done by `sim`. Perhaps this is not a real problem, and time and space -varying growth rates or carrying capacities could be done by handing the initialize function a list of raster objects.

`get_observations` documentation is a little short, but mainly just needs a good sentence or two upfront as the description. It all starts to make sense reading through the `@param` arguments. Could be interesting to work this function into the existing package vignette (this is not strictly necessary, but it might help highlight all the functionality the package has to offer).

Are the LICENSE and LICENSE.md both needed? They could be, but I’m just making sure.

Is it worth including a `seed` argument to the `sim` function? This is not strictly necessary, but for reproducibility it means the end user will be responsible for setting the seed before the `sim` function.

Otherwise it looks great. Really well-documented and useful package. It installs easily and passes R package checks. ran through all vignette code and played around with most (if not all) the functions in the package.

Check duration: 1m 45.6s 0 errors | 0 warnings | 1 note

> checking top-level files … NOTE Non-standard file/directory found at top level: `CODE_OF_CONDUCT.md`

Below is random code from me fiddling around with the package (included for completeness, but feel free to ignore). Some of the simulations take a very long time (e.g., if dispersal rate is really high or if landscape size gets large). If the simulation is initialized with all values of n1 (e.g., `n1_small`) set to 0, it does not error out that gracefully. It could easily be asked ‘why would the user do this?’, but probably best to cover as many bases with informative error messages?

    getSDSim <- function(n1=n1_small, K=K_small,
      r_sd=0, K_sd=0, timez=100, d=1){

      sim_data_01 <- initialise(
        n1_map = n1,
        K_map = K,
        r = log(2),
        r_sd=r_sd,
        K_sd=K_sd,
        rate = d
      )

      sim_result_01 <- sim(obj = sim_data_01, time = timez)
      plot(sim_result_01)
      return(sim_result_01)
    }

    # playing around with r_sd and seed-set behavior
    sim0 <- getSDSim(r_sd=0, timez=500)

    set.seed(123)
    sim1 <- getSDSim(r_sd=log(2), timez=500)
    set.seed(123)
    sim1b <- getSDSim(r_sd=log(2), timez=500)

    sim2 <- getSDSim(r_sd=log(4), timez=500)

    # working through some random simulations on constructed landscapes
    makeLandscape <- function(popSize=4, n=10, m=10){
      r <- terra::rast(matrix(popSize, nrow=n, ncol=m))
      crs(r) <- "+proj=lcc +lat_1=48 +lat_2=33 +lon_0=-100 +ellps=WGS84"
      return(r)
    }

    n1 <- makeLandscape(10,10,10)
    k1 <- n1
    values(k1) <- rep(100, 100)

    sim1a <- getSDSim(n1, k1, r_sd=0, d=0.001, timez=100)
    sim2a <- getSDSim(n1, k1, r_sd=1, d=0.001, timez=100)

    n0 <- makeLandscape(0,10,10)
    sim0start <- getSDSim(n0, k1, timez=10)

> Error in .as.raster.continuous(out, x, type) : out*r**a**n**g**e*\[2\] &gt; *o**u**t*range\[1\] is not TRUE
