+++ DESCRIPTION
@@ -4,7 +4,8 @@ Title: Mechanistic Simulation of Species Range Dynamics
 Version: 1.0.0
 Authors@R: c(
     person("Katarzyna", "Markowska", email = "katarzyna.markowska@amu.edu.pl", role = c("aut", "cre")),
-    person("Lechosław", "Kuczyński", email = "lechu@amu.edu.pl", role = "aut"))
+    person("Lechosław", "Kuczyński", email = "lechu@amu.edu.pl", role = "aut"),
+    person("Tad", "Dallas", comment = c(github = "@taddallas"), role = "rev"))
 Description: Species range dynamics simulation toolset.
 License: MIT + file LICENSE
 Imports:
