+++ R/get_observations.R
@@ -1,6 +1,9 @@
 #' Observation process
 #'
-#' This function simulates observation process.
+#' This function simulates observation process. It takes `sim_results` object
+#' generated by [`sim`] function and uses Virtual Ecologist approach on
+#' the `N_map` component of this object. As a result a `data.frame` with
+#' "observed" abundances is returned.
 #'
 #' @param sim_data  `sim_data` object from [`initialise`] containing simulation
 #' parameters
