+++ man/get_observations.Rd
@@ -81,7 +81,10 @@ estimated abundances, including observation error (if \code{sdlog > 0})
 and observer identifiers (if \code{type = "monitoring_based"})
 }
 \description{
-This function simulates observation process.
+This function simulates observation process. It takes \code{sim_results} object
+generated by \code{\link{sim}} function and uses Virtual Ecologist approach on
+the \code{N_map} component of this object. As a result a \code{data.frame} with
+"observed" abundances is returned.
 }
 \examples{
 \dontrun{
