+++ man/initialise.Rd
@@ -67,13 +67,13 @@ the standard deviation for a normal distribution around \code{r}
 
 \item{growth}{character vector of length 1; the name of a population growth
 function, either defined in \code{\link{growth}} or provided by
-the user (case-sensitive)}
+the user (case-sensitive, default \code{\link[=growth]{"gompertz"}})}
 
 \item{A}{numeric vector of length 1; strength of the Allee effect
 (see the \code{\link{growth}} function)}
 
 \item{dens_dep}{character vector of length 1 specifying if the probability
-of settling in a target square is (case-sensitive):
+of settling in a target square is (case-sensitive, default \code{"K2N"}):
 \itemize{
 \item{"none" - fully random,}
 \item{"K" - proportional to the carrying capacity of a target square,}
@@ -83,7 +83,7 @@ already present in a target square}
 }}
 
 \item{border}{character vector of length 1 defining how to deal
-with borders (case-sensitive):
+with borders (case-sensitive, default \code{"absorbing"}):
 \itemize{
 \item "absorbing" - individuals that disperse outside the study area
 are removed from the population
@@ -92,7 +92,7 @@ as targets for dispersal
 }}
 
 \item{kernel_fun}{character vector of length 1; name of a random number
-generation function defining a dispersal kernel (case-sensitive)}
+generation function defining a dispersal kernel (case-sensitive, default \code{"rexp"})}
 
 \item{...}{any parameters required by \code{kernel_fun}}
 
