### 2023-09-02 d9976d6
Minor changes to the description of VE
