+++ vignettes/rangr.Rmd
@@ -94,7 +94,6 @@ n1_small <- rast(system.file("input_maps/n1_small.tif", package = "rangr"))
 K_small <-  rast(system.file("input_maps/K_small.tif", package = "rangr"))
 ```
 
-
 Since both of these maps refer to the same virtual or in this case real environment (a small region located in northwestern Poland), they must have the same dimensions, resolution, geographical projection etc. The only differences between them may be the values they contain and the number of layers. You can use multiple layers in the carrying capacity map to create a dynamic environment during the simulation, but in this example, we will demonstrate a static environment.
 
 Let's take a closer look at them:
@@ -147,7 +146,7 @@ We have now set up the simulation environment along with the demographic and dis
 
 But first, let's see what other information `sim_data` contains and what happened behind the scenes during initialisation.
 
-First, let’s check the class of `sim_data` object:
+First, let's check the class of `sim_data` object:
 
 ```{r class_sim_data_01}
 class(sim_data_01)
@@ -238,9 +237,10 @@ plot(my_rast, c(1, 10, 25, 50))
 ```
 
 ## Virtual Ecologist
-If you want to sample your simulation results and mimic the real observation process you can use `get_observations()` function. Depending on type of observation that you choose (`type` argument), observation sites can be chosen randomly or be provided as a `data.frame`. Getting observation from simulated abundances may be done in order to test different methods of modelling species distribution and their ability to regenerate the original (in this case simulated) data. Other use case would be to determine the threshold number of study sites to reliably find out the species conservation status.
 
-Here we will demonstrate the most basic variant of `get_observations()` function. To to so we will need both the `sim_results` object and `sim_data` object used to generate the simulated data. We will randomly choose observation sites and with this variant we have to specify what proportion of all sites we want to conduct the observations on. Here we will choose 10% of sites so the `prop` argument is set to 0.1. There are two random observation processes available: one where observation sites stay the same throughout all time steps, and the other where at every time step observation sites are different. To see how the first version works, we set the `type` argument to `"random_one_layer"`.
+If you want to sample the abundance of your virtual species population and mimic the observation process, you can use the `get_observations()` function. Depending on the `type` argument, observation sites can be selected randomly or their coordinates can be provided as a `data.frame`.
+
+Here we will demonstrate the most basic variant of the `get_observations()` function. To do this we will need both the `sim_results` object and the `sim_data` object used to generate the simulated data. We will randomly select observation sites, and in this variant we need to specify what proportion of all sites we want to make observations on. Here we will select 10% of the sites, so the `prop` argument is set to 0.1. There are two random observation processes available: one where the observation sites are the same at all time steps, and the other where the observation sites are different at each time step. To see how the first version works, we set the `type` argument to `"random_one_layer"`.
 
 ```{r ve_01}
 set.seed(123)
@@ -253,15 +253,15 @@ sample_01 <- get_observations(
 str(sample_01)
 ```
 
-The returned `data.frame` has geographic coordinates of sampled observation sites, abundances and time steps from which the abundances are observed. To confirm that in each time steps the study sites stay the same and to get the coordinates of those sites we can use the following code:
+The returned `data.frame` contains the geodetic coordinates of the survey sites, the population abundances, and the time steps from which the abundances were observed. To confirm that the survey sites remain the same in each time step, and to obtain the coordinates of these sites, we can use the following code:
 
 ```{r ve_coords_01}
 unique(sample_01[c("x", "y")])
 ```
 
-Here we got 15 sites sampled.
+Here we sampled 15 sites.
 
[…]
