### 2023-09-04 popecol/rangr#2 @katarzynam-165
r_sd: demographic or environmental stochasticity 
Address the issue raised by @taddallas during [rOpenSci software peer review process](https://github.com/ropensci/software-review/issues/595):

> A small note, but the authors argue that r_sd is demographic stochasticity, but it is a time-varying growth rate applied to all cells equally (this may or may not be right). This seems like a measure of environmental stochasticity (e.g., there are good years and bad years). Perhaps future development of the package could include spatial variation in growth rates? Covariation between r and K (or the variance therein)? One way the authors could do this is to have the r argument take any number of forms (a raster the same size as n, a single value, etc.). It’s tough to do the change through time bits, as the initialize step sets up so much of the hyperparameters, but the simulation is done by sim. Perhaps this is not a real problem, and time and space -varying growth rates or carrying capacities could be done by handing the initialize function a list of raster objects.
