### 2023-09-04 457e57a
update get_observations documentation
