+++ R/get_observations.R
@@ -1,6 +1,6 @@
 #' Observation process
 #'
-#' This function simulates observation process. It takes `sim_results` object
+#' This function simulates an observation process. It takes `sim_results` object
 #' generated by [`sim`] function and uses Virtual Ecologist approach on
 #' the `N_map` component of this object. As a result a `data.frame` with
 #' "observed" abundances is returned.
