+++ NEWS.md
@@ -1,0 +1,6 @@
+# rangr 1.0.1 (2023-09-04)
+
+### MINOR IMPROVEMENTS
+
+-   Improved documentation for `get_observations`.
+-   Added functionality description for `get_observations` to the vignettes
