### 2023-09-05 #595 @katarzynam-165
@adamhsparks, we appreciate your ongoing efforts to secure a second reviewer for our package. We do have two potential candidates in mind who could be suitable, although we cannot guarantee their availability or willingness to participate in the review:
- [Anne-Kathleen Malchow](https://www.uni-potsdam.de/en/ibb-macroecology/team/anne-kathleen-malchow)
- [Casey Visintin](https://findanexpert.unimelb.edu.au/profile/666315-casey-visintin)
