### 2023-09-05 #595 @katarzynam-165
@taddallas, once again, thank you for taking the time to review our package, and for your valuable feedback. We greatly appreciate your efforts in evaluating our work.
1.	We added you to the DESCRIPTION file in the following way:` person("Tad", "Dallas", comment = c(github = "@taddallas"), role = "rev"))`. Please inform us if you'd like us to include your email or any additional information.
2.	We acknowledge that the `raster `package will no longer be supported soon. We've developed the code exclusively using the `terra `package and are actively working on optimizing it. Users should experience no difference in using our package.
3.	We've included the Code_of_Conduct.md file in .Rbuildignore as per your suggestion.
4.	We updated the documentation for the `initialize `function and included the default values for all relevant arguments.
5.	We appreciate your insight into the interpretation of `r_sd `as demographic vs. environmental stochasticity. We will explore this concept in the future – for now we added this as a [GitHub issue](https://github.com/popecol/rangr/issues/2).
6.	Documentation of the `get_observation` function was indeed lacking. We added some examples to the package vignette and expanded the help file a little.
7.	LICENSE and LICENSE.md are both needed and serve different purposes. More explanation can be found [here](https://r-pkgs.org/license.html#key-files).
8.	In our experience setting seed by the user before executing the function is more common and we find it more straightforward. Therefore, we've opted not to include a seed argument in the `sim `function.

Once again, we extend our gratitude for your thorough review. Please feel free to reach out if you have any further suggestions or inquiries.
