### 2023-09-28 #595 @katarzynam-165
@adamhsparks, I was wondering if there's been any progress in finding a second reviewer for `rangr`? Any updates would be greatly appreciated. Thank you!
