### 2023-11-01 #595 @TheAnalyticalEdge
## Package Review

- ☒ As the reviewer I confirm that there are no [conflicts of interest](https://devguide.ropensci.org/policies.html#coi) for me to review this work (If you are unsure whether you are in conflict, please speak to your editor _before_ starting your review).

---

### General Review

#### Documentation

- ☒ **A statement of need** clearly stating problems the software is designed to solve and its target audience in README
- ☒ **Installation instructions:** for the development version of package and any non-standard dependencies in README
- ☒ **Community guidelines** including contribution guidelines in the README or CONTRIBUTING
- ☒ The documentation is sufficient to enable general use of the package beyond one specific use case

#### Visualisation (where appropriate)

- Do visualisations aid the primary purposes of statistical interpretation of results?

Yes. I'd like to see ability to manipulate rows/cols in the plots. E.g.,

# generate visualisation
plot(sim_result_01,
     time_points = c(1, 10, 25, 50, 75, 100),
     template = sim_data_01$K_map,
     **nrow=3**
)

It will make it easier for reporting if plots can be put in to certain dimensions.

#### Package Design

- In relation to **Internal Design:** Could ranges of admissible input structures, or form(s) of output structures, be expanded to enhance inter-operability with other packages?

I'm worried this package still has warnings for dependancies on sp, which is now retired.

---

- ☒ **Packaging guidelines**: The package conforms to the rOpenSci packaging guidelines

Estimated hours spent reviewing: 3-4 hours.

- ☒ Should the author(s) deem it appropriate, I agree to be acknowledged as a package reviewer ("rev" role) in the package DESCRIPTION file.

Extra comments to the Author:

I see a clear need for this package in statistical ecology, and found it fairly easy to implement and use. Help files are all great and good to follow. Code in worked examples all works, etc.

In the get_observations() function, I wonder if extra options for sdlog can be provided. I believe this is some 'random noise' added to the observation to simulate the detectability effect. But often detectability is one-way (people count less animals than present) and for future, having different options in here would be good.

In the help files, contrasting ?initialise (good) and ?sim (less good), can the formatting of the input arguments be changed? the helpful arguments run off the screen which makes it annoying to read (picky of me, I know!) compared to ?initialise where  the arguments are all on separate lines.

There is a minor typo on the help page: 'and more suitable for _the_ virtual species'. Just after creating 'K_small_changing' object.

HTH :)
