### 2023-11-01 #595 @adamhsparks
Dear @katarzynam-165, we now have the second review complete, thank you @TheAnalyticalEdge! I do share the concern that both reviewers raised about the use of packages relying on {sp}, as this was officially retired last month. Can you please update the codebase to remove this dependency on {raster} completely by moving to {terra}?
