### 2023-11-01 #595 @katarzynam-165
@TheAnalyticalEdge thank you for your review and valuable suggestions! I appreciate your feedback.
@adamhsparks I'm currently attending a statistical course until the end of this week, but I'll make the necessary changes next week. I'll also address the reviewer's suggestions then. Sorry for the delay!
