+++ R/rangr-package.R
@@ -10,9 +10,6 @@
 #' @importFrom parallel parLapply
 #' @importFrom parallel parLapplyLB
 #' @importFrom pbapply pblapply
-#' @importFrom raster as.matrix
-#' @importFrom raster distanceFromPoints
-#' @importFrom raster raster
 #' @importFrom stats approx
 #' @importFrom stats getCall
 #' @importFrom stats quantile
@@ -22,6 +19,7 @@
 #' @importFrom stats rnorm
 #' @importFrom stats rpois
 #' @importFrom terra app
+#' @importFrom terra as.matrix
 #' @importFrom terra classify
 #' @importFrom terra compareGeom
 #' @importFrom terra crs
