+++ NAMESPACE
@@ -27,9 +27,6 @@ importFrom(parallel,clusterExport)
 importFrom(parallel,parLapply)
 importFrom(parallel,parLapplyLB)
 importFrom(pbapply,pblapply)
-importFrom(raster,as.matrix)
-importFrom(raster,distanceFromPoints)
-importFrom(raster,raster)
 importFrom(stats,approx)
 importFrom(stats,getCall)
 importFrom(stats,quantile)
@@ -43,6 +40,7 @@ importFrom(terra,"nlyr<-")
 importFrom(terra,"res<-")
 importFrom(terra,"values<-")
 importFrom(terra,app)
+importFrom(terra,as.matrix)
 importFrom(terra,classify)
 importFrom(terra,compareGeom)
 importFrom(terra,crs)
