### 2023-11-27 #595 @adamhsparks
Hi @katarzynam-165, just checking on how you are going with the changes to {rangr}. I'm wrapping up for the year and will be on leave for several weeks before starting a new position in the new year. Is it possible for us to finalise this in the next two weeks?
