+++ R/plot.sim_results.R
@@ -36,6 +36,12 @@
 #' sim_res <- sim(sim_data, time = 10)
 #' plot(sim_res)
 #'
+#' # plot specific area
+#' plot(sim_res, xlim = c(4, 10), ylim = c(0, 10))
+#' plot(sim_res, ext = c(4, 10, 0, 10))
+#' plot(sim_res, template = n1_small, ext = c(274000, 280000, 610000, 620000))
+#'
+#'
 #' @srrstats {G1.4} uses roxygen documentation
 #' @srrstats {G2.0a} documented lengths expectation
 #' @srrstats {G2.1a, G2.3, G2.3b, SP2.6} documented types expectation
