### 2023-11-28 #595 @katarzynam-165
Hi @adamhsparks, great timing! I've successfully implemented all the necessary changes in `rangr`. I appreciate your patience and apologize for any delays. I'll be posting the review response shortly. Wishing you a pleasant leave and success in your new role! :)
