### 2023-11-28 #595 @katarzynam-165
@TheAnalyticalEdge, I'm grateful for the time and effort you dedicated to provide feedback. I've addressed several of the points in your review, and I hope these changes are more in line with your expectations:
1. **Dependencies**: We've removed the sp dependencies.
2. **Plotting Function**: Regarding the plot function, I’m no sure if I understand your comment correctly. If your intention is to selectively plot specific areas, you can absolutely do this . All parameters unused by plot.sim_results are seamlessly passed to the plot function within the terra package. It allows for the following code to be used:
```
plot(sim_res, xlim = c(4, 10), ylim = c(0, 10))
plot(sim_res, ext = c(4, 10, 0, 10))
plot(sim_res, template = n1_small, ext = c(274000, 280000, 610000, 620000))
```
These examples have been added to [the help file](https://popecol.github.io/rangr/reference/plot.sim_results.html) for clarity.

3.  **Observation error distribution**: In addition to the already available log-normal distribution (which can accommodate both commission and omission errors), we have added the binomial distribution (to emulate incomplete detectability). The user can now specify the type of distribution (`rlnorm `or `rbinom`) and a parameter for this distribution (log-sd or probability of success).
4. **Updated sim function documentation**: We've modified the sim function as illustrated in the attached screenshot. I hope it is less annoying now :)
![image](https://github.com/ropensci/software-review/assets/37637319/0d1e84da-f22d-4ddf-bcf2-cb8fa35a7b57)

5. **Typo Correction**: The typo is fixed. Thank you for catching that!

I genuinely hope these adjustments meet your expectations. If you have any further suggestions or concerns, please do not hesitate to let me know.
