### 2023-11-28 #595 @adamhsparks
Thank you, @katarzynam-165, for your thoughtful revisions.

@taddallas, @TheAnalyticalEdge could you please let me know if you have any further suggestions or if you're happy with the changes as they are?
