+++ README.Rmd
@@ -34,7 +34,7 @@ library(rangr)
 The **rangr** package is designed to simulate a species range dynamics. This new tool mimics the essential processes that shape population numbers and spatial distribution: local dynamics and dispersal. Simulations can be run in a spatially explicit and dynamic environment, facilitating population projections in response to climate or land-use changes. By using different sampling schemes and observational error distributions, the structure of the original survey data can be reproduced, or pure random sampling can be mimicked.
 
 The study is supported by the National Science Centre, Poland, grant no.
-2018/29/B/NZ8/00066.
+2018/29/B/NZ8/00066 and the Poznań Supercomputing and Networking Centre (grant no. 403).
 
 # Installation
 
