### 2024-01-08 #595 @adamhsparks
@taddallas, @TheAnalyticalEdge, Happy New Year!

Can you please approve the changes made to {rangr} or provide further suggestions for @katarzynam-165 to implement so that we can finalise this review? Thanks!
