### 2024-01-13 #595 @adamhsparks
@katarzynam-165, I've detected a few issues with the documentation that should be addressed before we proceed.

Please rerun the regular R package checks, _e.g._, `devtools::check()`, and take note of errors regarding Unnecessary braces that occur in initialise.R's ROxygen header and correct them.

Also, I noted that your documentation titles are incorrectly formatted. They should be Title Case, currently you have sentence case.
