+++ R/get_observations.R
@@ -1,4 +1,4 @@
-#' Observation process
+#' Observation Process
 #'
 #' This function simulates an observation process. It takes `sim_results` object
 #' generated by [`sim`] function and uses Virtual Ecologist approach on
@@ -204,7 +204,7 @@ get_observations <- function(
 
 
 
-#' Random sampling
+#' Random Sampling
 #'
 #' `get_observations` calls this function if sampling type equals to
 #' "random_one_layer" or "random_all_layers".
@@ -272,7 +272,7 @@ get_observations_random <- function(N_rast, type, prop = 0.1) {
 }
 
 
-#' Sampling based on given data.frame
+#' Sampling Based On Given Data
 #'
 #' [get_observations] calls this function if sampling type
 #' equals to "from_data".
@@ -325,7 +325,7 @@ get_observations_from_data <- function(N_rast, points) {
 
 
 
-#' Sampling that mimics a real survey programmes
+#' Sampling That Mimics A Real Survey Programmes
 #'
 #' [get_observations] calls this function if sampling type
 #' equals to "monitoring_based".
