+++ R/subset.sim_results.R
@@ -19,7 +19,7 @@
 #' @method subset sim_results
 #'
 #' @examples
-#'
+#' \dontrun{
 #' # data preparation
 #' library(terra)
 #'
@@ -39,6 +39,7 @@
 #'
 #' sim_results_cropped <- subset(sim_results, time_points = c(1:2))
 #' summary(sim_results_cropped)
+#' }
 #'
 #' @srrstats {G1.4} uses roxygen documentation
 #' @srrstats {G2.0a} documented lengths expectation
