+++ R/plot.sim_results.R
@@ -21,7 +21,7 @@
 #' @method plot sim_results
 #'
 #' @examples
-#'
+#' \dontrun{
 #' library(terra)
 #'
 #' n1_small <- rast(system.file("input_maps/n1_small.tif", package = "rangr"))
@@ -40,7 +40,7 @@
 #' plot(sim_res, xlim = c(4, 10), ylim = c(0, 10))
 #' plot(sim_res, ext = c(4, 10, 0, 10))
 #' plot(sim_res, template = n1_small, ext = c(274000, 280000, 610000, 620000))
-#'
+#' }
 #'
 #' @srrstats {G1.4} uses roxygen documentation
 #' @srrstats {G2.0a} documented lengths expectation
