### 2024-01-18 #595 @katarzynam-165
@adamhsparks, I updated the documentation titles to be Title Case, but I'm unable to replicate erros that you mentioned. I've run `devtools::check` and `pkgcheck::pkgcheck` and none of them returned any errors or warning. Can you please provide me with a specific function that you used? Or maybe a screen shot?
