+++ man/plot.sim_results.Rd
@@ -32,7 +32,7 @@ as the length of \code{time_points} parameter
 Plots abundances obtained during simulation.
 }
 \examples{
-
+\dontrun{
 library(terra)
 
 n1_small <- rast(system.file("input_maps/n1_small.tif", package = "rangr"))
@@ -51,6 +51,6 @@ plot(sim_res)
 plot(sim_res, xlim = c(4, 10), ylim = c(0, 10))
 plot(sim_res, ext = c(4, 10, 0, 10))
 plot(sim_res, template = n1_small, ext = c(274000, 280000, 610000, 620000))
-
+}
 
 }
