### 2024-01-20 #595 @adamhsparks
This is what I get.

```r
> devtools::document()
ℹ Updating rangr documentation
ℹ Loading rangr
> devtools::check()
══ Documenting ══════════════════════════════════════════════════════════════════════════════════════
ℹ Updating rangr documentation
ℹ Loading rangr

══ Building ═════════════════════════════════════════════════════════════════════════════════════════
Setting env vars:
• CFLAGS    : -Wall -pedantic -fdiagnostics-color=always
• CXXFLAGS  : -Wall -pedantic -fdiagnostics-color=always
• CXX11FLAGS: -Wall -pedantic -fdiagnostics-color=always
• CXX14FLAGS: -Wall -pedantic -fdiagnostics-color=always
• CXX17FLAGS: -Wall -pedantic -fdiagnostics-color=always
• CXX20FLAGS: -Wall -pedantic -fdiagnostics-color=always
── R CMD build ──────────────────────────────────────────────────────────────────────────────────────
✔  checking for file ‘/Users/adamsparks/Developer/GitHub/rOpenSci/rangr/DESCRIPTION’ ...
─  preparing ‘rangr’:
✔  creating vignettes (13.7s)
─  building ‘rangr_1.0.2.tar.gz’

══ Checking ═════════════════════════════════════════════════════════════════════════════════════════
Setting env vars:
• _R_CHECK_CRAN_INCOMING_REMOTE_               : FALSE
• _R_CHECK_CRAN_INCOMING_                      : FALSE
• _R_CHECK_FORCE_SUGGESTS_                     : FALSE
• _R_CHECK_PACKAGES_USED_IGNORE_UNUSED_IMPORTS_: FALSE
• NOT_CRAN                                     : true
── R CMD check ──────────────────────────────────────────────────────────────────────────────────────
─  using log directory ‘/private/var/folders/ch/8fqkzddj1kj_qb5ddfdd3p1w0000gn/T/RtmpTAXSaq/file21206d5d0f6/rangr.Rcheck’
─  using R version 4.3.2 (2023-10-31)
─  using platform: aarch64-apple-darwin20 (64-bit)
─  R was compiled by
       Apple clang version 14.0.0 (clang-1400.0.29.202)
       GNU Fortran (GCC) 12.2.0
─  running under: macOS Sonoma 14.2.1
─  using session charset: UTF-8
─  using options ‘--no-manual --as-cran’
✔  checking for file ‘rangr/DESCRIPTION’
─  checking extension type ... Package
─  this is package ‘rangr’ version ‘1.0.2’
─  package encoding: UTF-8
✔  checking package namespace information
✔  checking package dependencies (719ms)
✔  checking if this is a source package ...
✔  checking if there is a namespace
✔  checking for executable files ...
✔  checking for hidden files and directories
✔  checking for portable file names
✔  checking for sufficient/correct file permissions
─  checking whether package ‘rangr’ can be installed ... [5s/5s] OK (4.9s)
✔  checking installed package size
✔  checking package directory
✔  checking for future file timestamps (566ms)
✔  checking ‘build’ directory
✔  checking top-level files
✔  checking for left-over files
✔  checking index information ...
✔  checking package subdirectories ...
✔  checking R files for non-ASCII characters ...
✔  checking R files for syntax errors ...
─  checking whether the package can be loaded ... [1s/1s] OK (1.4s)
─  checking whether the package can be loaded with stated dependencies ... [1s/1s] OK (1.4s)
─  checking whether the package can be unloaded cleanly ... [1s/1s] OK (1.4s)
─  checking whether the namespace can be loaded with stated dependencies ... [1s/1s] OK (1.4s)
─  checking whether the namespace can be unloaded cleanly ... [1s/1s] OK (1.4s)
─  checking loading without being on the library search path ... [1s/1s] OK (1.4s)
✔  checking dependencies in R code (2.8s)
✔  checking S3 generic/method consistency (1.4s)
✔  checking replacement functions (1.3s)
✔  checking foreign function calls (1.4s)
─  checking R code for possible problems ... [7s/7s] OK (7s)
─  checking Rd files ... [0s/0s] WARNING
   checkRd: (-3) disp.Rd:38: Unnecessary braces at ‘{"none" - fully random,}’
   checkRd: (-3) disp.Rd:39: Unnecessary braces at ‘{"K" - proportional to the carrying capacity of a target square,}’
   checkRd: (-3) disp.Rd:40-42: Unnecessary braces at ‘{"K2N" - density-dependent, i.e. proportional to the ratio of
   carrying capacity of a target square to the number of individuals
   already present in a target square}’
   checkRd: (-3) get_observations.Rd:54-58: Unnecessary braces at ‘{\code{prop}

   numeric vector of length 1; proportion of cells to be sampled
   (default \code{prop = 0.1});
   used when \verb{type = "random_one_layer" or "random_all_layers"},}’
   checkRd: (-3) get_observations.Rd:60-64: Unnecessary braces at ‘{\code{points}

   \code{data.frame} or \code{matrix} with 3 named numeric columns ("x", "y"
   and "time_step") containing coordinates and time steps from which
   observations should be obtained; used when \code{type = "from_data"},}’
   checkRd: (-3) get_observations.Rd:66-69: Unnecessary braces at ‘{\code{cells_coords}

   \code{data.frame} or \code{matrix} with 2 named columns: "x" and "y"; survey plots
   coordinates; used when \code{type = "monitoring_based"}}’
   checkRd: (-3) get_observations.Rd:71-77: Unnecessary braces at ‘{\code{prob}

   numeric vector of length 1; a parameter defining the shape of -
   \code{\link[stats:Geometric]{rgeom}} distribution - it defines whether an observation
   will be made by the same observer for several years and whether it
   will not be made at all (default \code{prob = 0.3});
   used when \code{type = "monitoring_based"}}’
   checkRd: (-3) get_observations.Rd:79-83: Unnecessary braces at ‘{\code{progress_bar}

   logical vector of length 1; determines if progress bar for observational
   process should be displayed (default \code{progress_bar = FALSE});
   used when \code{type = "monitoring_based"}}’
   checkRd: (-3) initialise.Rd:78: Unnecessary braces at ‘{"none" - fully random,}’
   checkRd: (-3) initialise.Rd:79: Unnecessary braces at ‘{"K" - proportional to the carrying capacity of a target square,}’
   checkRd: (-3) initialise.Rd:80-82: Unnecessary braces at ‘{"K2N" - density-dependent, i.e. proportional to the ratio of
   carrying capacity of a target square to the number of individuals
   already present in a target square}’
   checkRd: (-3) sim.Rd:41: Unnecessary braces at ‘{\code{extinction} - \code{TRUE} if population is extinct or \code{FALSE} otherwise}’
   checkRd: (-3) sim.Rd:42-43: Unnecessary braces at ‘{\code{simulated_time} - number of simulated time steps without
   the burn-in ones}’
   checkRd: (-3) sim.Rd:44-46: Unnecessary braces at ‘{\code{N_map} - 3-dimensional array representing spatio-temporal
   variability in population numbers. The first two dimensions correspond to
   the spatial aspect of the output and the third dimension represents time.}’
✔  checking Rd metadata ...
✔  checking Rd line widths ...
✔  checking Rd cross-references ...
✔  checking for missing documentation entries (1.4s)
✔  checking for code/documentation mismatches (4.1s)
✔  checking Rd \usage sections (1.5s)
✔  checking Rd contents ...
✔  checking for unstated dependencies in examples ...
✔  checking contents of ‘data’ directory
─  checking data for non-ASCII characters ... [0s/0s] OK
✔  checking LazyData
✔  checking data for ASCII and uncompressed saves ...
✔  checking installed files from ‘inst/doc’
✔  checking files in ‘vignettes’ ...
─  checking examples ... [5s/5s] OK (5.3s)
✔  checking for unstated dependencies in ‘tests’ ...
─  checking tests ...
    [11s/11s] OKthat.R’
   * checking for unstated dependencies in vignettes ... OK
   * checking package vignettes in ‘inst/doc’ ... OK
   * checking re-building of vignette outputs ... [8s/8s] OK
   * checking for non-standard things in the check directory ... OK
   * checking for detritus in the temp directory ... OK
   * DONE

   Status: 1 WARNING
   See
     ‘/private/var/folders/ch/8fqkzddj1kj_qb5ddfdd3p1w0000gn/T/RtmpTAXSaq/file21206d5d0f6/rangr.Rcheck/00check.log’
   for details.

── R CMD check results ───────────────────────────────────────────────────────────── rangr 1.0.2 ────
Duration: 1m 2.8s

❯ checking Rd files ... [0s/0s] WARNING
  checkRd: (-3) disp.Rd:38: Unnecessary braces at ‘{"none" - fully random,}’
  checkRd: (-3) disp.Rd:39: Unnecessary braces at ‘{"K" - proportional to the carrying capacity of a target square,}’
  checkRd: (-3) disp.Rd:40-42: Unnecessary braces at ‘{"K2N" - density-dependent, i.e. proportional to the ratio of
  carrying capacity of a target square to the number of individuals
  already present in a target square}’
  checkRd: (-3) get_observations.Rd:54-58: Unnecessary braces at ‘{\code{prop}

  numeric vector of length 1; proportion of cells to be sampled
  (default \code{prop = 0.1});
  used when \verb{type = "random_one_layer" or "random_all_layers"},}’
  checkRd: (-3) get_observations.Rd:60-64: Unnecessary braces at ‘{\code{points}

  \code{data.frame} or \code{matrix} with 3 named numeric columns ("x", "y"
  and "time_step") containing coordinates and time steps from which
  observations should be obtained; used when \code{type = "from_data"},}’
  checkRd: (-3) get_observations.Rd:66-69: Unnecessary braces at ‘{\code{cells_coords}

  \code{data.frame} or \code{matrix} with 2 named columns: "x" and "y"; survey plots
  coordinates; used when \code{type = "monitoring_based"}}’
  checkRd: (-3) get_observations.Rd:71-77: Unnecessary braces at ‘{\code{prob}

  numeric vector of length 1; a parameter defining the shape of -
  \code{\link[stats:Geometric]{rgeom}} distribution - it defines whether an observation
  will be made by the same observer for several years and whether it
  will not be made at all (default \code{prob = 0.3});
  used when \code{type = "monitoring_based"}}’
  checkRd: (-3) get_observations.Rd:79-83: Unnecessary braces at ‘{\code{progress_bar}

  logical vector of length 1; determines if progress bar for observational
  process should be displayed (default \code{progress_bar = FALSE});
  used when \code{type = "monitoring_based"}}’
  checkRd: (-3) initialise.Rd:78: Unnecessary braces at ‘{"none" - fully random,}’
  checkRd: (-3) initialise.Rd:79: Unnecessary braces at ‘{"K" - proportional to the carrying capacity of a target square,}’
  checkRd: (-3) initialise.Rd:80-82: Unnecessary braces at ‘{"K2N" - density-dependent, i.e. proportional to the ratio of
  carrying capacity of a target square to the number of individuals
  already present in a target square}’
  checkRd: (-3) sim.Rd:41: Unnecessary braces at ‘{\code{extinction} - \code{TRUE} if population is extinct or \code{FALSE} otherwise}’
  checkRd: (-3) sim.Rd:42-43: Unnecessary braces at ‘{\code{simulated_time} - number of simulated time steps without
  the burn-in ones}’
  checkRd: (-3) sim.Rd:44-46: Unnecessary braces at ‘{\code{N_map} - 3-dimensional array representing spatio-temporal
  variability in population numbers. The first two dimensions correspond to
  the spatial aspect of the output and the third dimension represents time.}’

0 errors ✔ | 1 warning ✖ | 0 notes ✔
```

System information: R 4.3.2, MacOS 14.2.1, {roxygen2} version 7.3.0
