### 2024-01-20 97b068b
remove unnecessary braces from documentation
