+++ R/initialise.R
@@ -44,11 +44,11 @@
 #' @param dens_dep character vector of length 1 specifying if the probability
 #' of settling in a target square is (case-sensitive, default `"K2N"`):
 #' \itemize{
-#'   \item{"none" - fully random,}
-#'   \item{"K" - proportional to the carrying capacity of a target square,}
-#'   \item{"K2N" - density-dependent, i.e. proportional to the ratio of
+#'   \item "none" - fully random,
+#'   \item "K" - proportional to the carrying capacity of a target square,
+#'   \item "K2N" - density-dependent, i.e. proportional to the ratio of
 #'   carrying capacity of a target square to the number of individuals
-#'   already present in a target square}
+#'   already present in a target square
 #' }
 #' @param border character vector of length 1 defining how to deal
 #' with borders (case-sensitive, default `"absorbing"`):
