### 2024-01-20 #595 @katarzynam-165
I've managed to replicate those warnings using `tools::checkRd()` and hopefully eliminated any unnecessary braces. Please let me know if everything is alright on your end too.
