### 2024-01-21 #595 @adamhsparks
Yes, you've corrected these issues.

One minor thing, you use `print()` in your Vignette. It isn't necessary to use that function directly, you can just call the object, _e.g._ line 374 `r > print(sim_data_01)` should just be `r > sim_data_01`. Minor, but inexperienced users might find it confusing.

I'll ask the bot to approve the package now. @taddallas and @TheAnalyticalEdge, thank you for your work here. @katarzynam-165, congratulations and thank you for your submission to rOpenSci!
