### 2024-01-21 #595 @ropensci-review-bot
Approved! Thanks @katarzynam-165 for submitting and taddallas, @taddallas, @TheAnalyticalEdge for your reviews! :grin:
